added pathname field to mountain and camera.
decided to use a separate field for pathname instead of generating
it from the name,etc. This way name will be 'display name' that I can
change without having to rename paths in the image folder, which would
be a pain.

I added necessary changes for this in the SQL, model, queries, and
tasks.

I also tweaked the ScheduleScrapes task a tiny bit.